requirements.txt
alembic==1.13.1
amqp==5.2.0
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
cachelib==0.10.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
email-validator==2.1.1
Flask
Flask-Caching==2.1.0
Flask-Limiter==3.4.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate
Flask-Moment
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.49.0
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==23.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
limits==4.0.1
Mako==1.3.3
Markdown==3.5.2
MarkupSafe==2.1.5
ordered-set==4.1.0
packaging==23.2
Pillow==10.2.0
pluggy==1.4.0
protobuf==4.25.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
redis==5.0.2
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.27
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
Werkzeug==3.0.1
WTForms==3.1.2
WTForms-SQLAlchemy==0.2.1
Flask-HTTPAuth==4.8.1
Flask-CLI==0.4.0
web3==6.13.0
Flask-Swagger-UI==4.13.0
Flask-Moment==1.1.0
pytest-flask==1.3.0

Flask-WTF
fonttools
frozenlist
greenlet
gunicorn
hexbytes
idna
iniconfig
itsdangerous
Jinja2
jmespath
jsonschema
jsonschema-specifications
kombu
limits
lru-dict
Mako
Markdown
markdown-it-py
MarkupSafe
mccabe
mdurl
multidict
mypy-extensions
ordered-set
packaging
parsimonious
pathspec
Pillow
platformdirs
pluggy
prompt_toolkit
propcache
protobuf
psycopg2-binary
pycodestyle
pycparser
pycryptodome
pydantic
pydantic_core
pydyf
pyflakes
Pygments
PyJWT
pyphen
pytest
pytest-flask
python-dateutil
python-dotenv
redis
referencing
regex
requests
rich
rlp
s3transfer
six
SQLAlchemy
tinycss2
tinyhtml5
tomli
toolz
typing-inspect
typing_extensions
tzdata
urllib3
vine
wcwidth
weasyprint
web3
webencodings
websockets
Werkzeug
wrapt
WTForms
WTForms-SQLAlchemy
yarl
zopfli